+++ autopush/db.py
@@ -1,6 +1,7 @@
 """Database Interaction"""
 from __future__ import absolute_import
 
+import datetime
 import logging
 import time
 import uuid
@@ -19,6 +20,51 @@ from boto.dynamodb2.types import NUMBER
 
 
 log = logging.getLogger(__file__)
+
+
+def get_month(delta=0):
+    """Basic helper function to get a datetime.date object iterations months
+    ahead/behind of now."""
+    new = last = datetime.date.today()
+    # Move until we hit a new month, this avoids having to manually
+    # check year changes as we push forward or backward since the Python
+    # timedelta math handles it for us
+    for _ in range(abs(delta)):
+        while new.month == last.month:
+            if delta < 0:
+                new -= datetime.timedelta(days=14)
+            else:
+                new += datetime.timedelta(days=14)
+        last = new
+    return new
+
+
+def make_rotating_tablename(prefix, delta=0):
+    """Creates a tablename for table rotation based on a prefix with a given
+    month delta."""
+    date = get_month(delta=delta)
+    return "{}_{}_{}".format(prefix, date.year, date.month)
+
+
+def create_rotating_message_table(prefix="message", read_throughput=5,
+                                  write_throughput=5, delta=0):
+    """Create a new message table for webpush style message storage"""
+    tablename = make_rotating_tablename(prefix, delta)
+    return Table.create(tablename,
+                        schema=[HashKey("uaid"),
+                                RangeKey("chidmessageid")],
+                        throughput=dict(read=read_throughput,
+                                        write=write_throughput),
+                        )
+
+
+def get_rotating_message_table(prefix="message", delta=0):
+    """Gets the message table for the current month.
+
+    This requires the table to already exist or errors will occur.
+
+    """
+    return Table(make_rotating_tablename(prefix, delta))
 
 
 def create_router_table(tablename="router", read_throughput=5,
@@ -266,9 +312,18 @@ class Message(object):
         try:
             result = self.table.get_item(consistent=True, uaid=uaid,
                                          chidmessageid=" ")
-            return result["chids"] or set([])
+            return (True, result["chids"] or set([]))
         except ItemNotFound:
-            return set([])
+            return False, set([])
+
+    @track_provisioned
+    def save_channels(self, uaid, channels):
+        """Save out a set of channels"""
+        self.table.put_item(data=dict(
+            uaid=uaid,
+            chidmessageid=" ",
+            chids=channels
+        ))
 
     @track_provisioned
     def store_message(self, uaid, channel_id, message_id, ttl, data=None,
@@ -483,6 +538,21 @@ class Router(object):
     @track_provisioned
     def drop_user(self, uaid):
         return self.table.delete_item(uaid=uaid)
+
+    @track_provisioned
+    def update_message_month(self, uaid, month):
+        """Update the route tables current_message_month"""
+        conn = self.table.connection
+        db_key = self.encode({"uaid": uaid})
+        expr = "SET current_month=:curmonth"
+        expr_values = self.encode({":curmonth": month})
+        conn.update_item(
+            self.table.table_name,
+            db_key,
+            update_expression=expr,
+            expression_attribute_values=expr_values,
+        )
+        return True
 
     @track_provisioned
     def clear_node(self, item):