+++ autopush/settings.py
@@ -1,14 +1,22 @@
 """Autopush Settings Object and Setup"""
+import datetime
 import socket
 
 from cryptography.fernet import Fernet, MultiFernet
 from twisted.internet import reactor
+from twisted.internet.defer import (
+    inlineCallbacks,
+    returnValue,
+)
+from twisted.internet.threads import deferToThread
+from twisted.python import log
 from twisted.web.client import Agent, HTTPConnectionPool
 
 from autopush.db import (
     get_router_table,
     get_storage_table,
-    get_message_table,
+    get_rotating_message_table,
+    make_rotating_tablename,
     preflight_check,
     Storage,
     Router,
@@ -136,15 +144,21 @@ class AutopushSettings(object):
         self.router_table = get_router_table(router_tablename,
                                              router_read_throughput,
                                              router_write_throughput)
-        self.storage_table = get_storage_table(storage_tablename,
-                                               storage_read_throughput,
-                                               storage_write_throughput)
-        self.message_table = get_message_table(message_tablename,
-                                               message_read_throughput,
-                                               message_write_throughput)
+        self.storage_table = get_storage_table(
+            storage_tablename,
+            storage_read_throughput,
+            storage_write_throughput)
+        self.message_table = get_rotating_message_table(
+            message_tablename)
+        self._message_prefix = message_tablename
         self.storage = Storage(self.storage_table, self.metrics)
         self.router = Router(self.router_table, self.metrics)
-        self.message = Message(self.message_table, self.metrics)
+
+        # Used to determine whether a connection is out of date with current
+        # db objects
+        self.current_msg_month = make_rotating_tablename(self._message_prefix)
+        self.current_month = datetime.date.today().month
+        self.create_initial_message_tables()
 
         # Run preflight check
         preflight_check(self.storage, self.router)
@@ -171,6 +185,61 @@ class AutopushSettings(object):
         self.env = env
 
         self.hello_timeout = hello_timeout
+
+    @property
+    def message(self):
+        """Property that access the current message table"""
+        return self.message_tables[self.current_msg_month]
+
+    @message.setter
+    def message(self, value):
+        """Setter to set the current message table"""
+        self.message_tables[self.current_msg_month] = value
+
+    def create_initial_message_tables(self):
+        """Initializes a dict of the initial rotating messages tables.
+
+        An entry for last months table, and an entry for this months table.
+
+        """
+        last_month = get_rotating_message_table(self._message_prefix, -1)
+        this_month = get_rotating_message_table(self._message_prefix)
+        self.message_tables = {
+            last_month.table_name: Message(last_month, self.metrics),
+            this_month.table_name: Message(this_month, self.metrics),
+        }
+
+    @inlineCallbacks
+    def update_rotating_tables(self):
+        """This method is intended to be tasked to run periodically off the
+        twisted event hub to rotate tables.
+
+        When today is a new month from yesterday, then we swap out all the
+        table objects on the settings object.
+
+        """
+        today = datetime.date.today()
+        if today.month == self.current_month:
+            # No change in month, we're fine.
+            returnValue(False)
+
+        # Get tables for the new month, and verify they exist before we try to
+        # switch over
+        message_table = get_rotating_message_table(self._message_prefix)
+
+        try:
+            yield deferToThread(message_table.describe)
+        except Exception:
+            tblname = make_rotating_tablename(self._message_prefix)
+            log.err("Unable to locate new message table: %s" % tblname)
+            returnValue(False)
+
+        # Both tables found, safe to switch-over
+        self.current_month = today.month
+        self.current_msg_month = message_table.table_name
+        self.message_tables[self.current_msg_month] = \
+            Message(message_table, self.metrics)
+        returnValue(True)
 
     def update(self, **kwargs):
         """Update the arguments, if a ``crypto_key`` is in kwargs then the